requirements.txt
Django>=5.1,<6.0
anthropic>=0.40.0
GitPython>=3.1.40
httpx>=0.27.0
Pillow>=10.2.0
beautifulsoup4>=4.12.0
markdownify>=0.13.0
gunicorn>=22.0.0
whitenoise>=6.7.0
python-dotenv>=1.0.0
PyMuPDF>=1.24.0
PyYAML>=6.0
dj-database-url>=2.1.0
psycopg2-binary>=2.9.0
